automacao_financeira_github/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.15
Flask-Migrate==4.0.4
psycopg2-binary==2.9.3
PyJWT==2.3.0
cryptography==36.0.1
PyPDF2==2.10.5
pandas==1.3.3
numpy==1.21.2
python-dateutil==2.8.2
Pillow==8.3.2

